thebaker-ai/requirements.txt
pandas
numpy
xgboost
torch
fastapi
uvicorn
langgraph
pydantic
google-generativeai
pytest
httpx
